monitorium/requirements.txt
# Cloud Functions runtime dependencies.
# Installed by `gcloud functions deploy` and used for local ingestion dev.
# Do NOT add pyspark here — Dataproc Serverless provides it.

functions-framework
google-cloud-storage
google-cloud-dataproc
requests
beautifulsoup4
lxml
yfinance
python-dotenv
curl_cffi
